Add line/branch coverage testing.
In Python, one of the major classes of errors (since no compilation
step exists) is simple typos of identifiers. Running coverage to
verify that each line is touched finds 99% of these typos.

This adds a simple test to run coverage over the "steve" package, and
achieves 100% for the primary package, but 0% of the "steve.vtypes"
subpackage (but will, shortly).