Erroneous use of strip() can cause hasVoted to return False on issues that have been voted on
As the intent seems to be removing the .json ending from the issue ID, let's be strict about it and instead remove the last five chars if they are `.json`. Otherwise, issue IDs that do not conform to hexadecimals will have their j, s, o, and n bits cut out.