Fix determination of the issue ID.
The Issue ID is always the second segment of the path. Extract this
early, which fixes a bug for the "delete" action which tests ISSUE
before it (never) gets set.

(because action=="delete" fails, the ISSUE name is never tested;
basically, nobody has tried deleting an issue via the API)