Revamp schema for multiple elections in one sqlite database.
Rather than one election per database, we're placing an entire
install's elections into one. This enables a user to easily query for
"all elections I can vote in" (eg. for a dashboard).

In the previous schema, the set of voters for the single election is
defined as the rows in the "person" databse. With multiple elections,
then we create a "mayvote" table to specify which Issues a Person may
vote upon (where an Issue is related to a specific Election).

"elections" table:
- add CHECKs for columns: eid, salt, opened_key, closed

"issues" table:
- add CHECK for the iid column
- remove salt column; no longer needed

"person" table:
- remove salt column; no longer needed

New "mayvote" table:
- pid/iid create the needed pairing of who may vote on what
- salt is used in a couple contexts; add a CHECK to its value

"votes" table:
- remove person_token and issue_token. Instead, use a single
  vote_token computed from the pid/iid and a salt.
- remove the salt column; the vote's related mayvote.salt is used via
  key-stretching to construct the cipher key.
- rename token to ciphertext for typical naming practive
- revise the two indexes into a single one on the vote_token column