Initial draft of state workflow for an Election.
Display a state diagrom from Editable to Open to Closed. Provide
buttons to advance the state of the Election. Confirmation dialogs
since these actions are irreversible.

TBD: hook the buttons into server interactions to actually perform the
workflow. This draft is client-side only.

Note: this is really janky. CSS outside of <head/> and some JS after
the final </html>. It works in Chrome, but this is not ideal.
Committing for now, to get the content in here, with future refinement.