Turn the crank on add/edit/delete of issues.
Turn deletion of an issue into a GET rather than a DELETE. This breaks
HTTP idioms, but I have not found a way to perform a DELETE (via
fetch() in the browser) and properly handle a 303 redirect (it can
work, but does a double-fetch of the Location target)

Iterate the add/edit endpoints to fetch the form data.

do_delete_issue_endpoint() now calls .delete_issue() to actually
delete the issue from the Election.

add flash messages to the issue-handling endpoints.

fix the JS in manage.ezt for the dynamic add/edit modal, and add an
id= to the delete buttons for deleteIssue() to find.