Focus on form submission via POST.
The use of fetch() to send DELETE created a pit of despair. The
fetch() function does not properly follow 303 redirects after the
operation completes. That led to trying alternatives, and then to GET
as a state-altering invocation (yes, bad, but trying to get 303 to
work). Meanwhile, trying to insert a confirm() call on these state-
altering features.

Answer: switch to a form, rather than confirm(). This allows using
Bootstrap for better UX alignment, and that form will send a POST
where the browser understands the 303 response and properly follows to
the desired Location.

* use a modal for confirming delete of an issue, not a confirm()
* for add/edit of an issue, switch to a form POST, not a fetch()

note: small TBD tweak: add csrf_token to basic template data