Split add/edit of issues.
We should not have conflated the method to add issues with that of
editing issues. These are now separate methods, and separate cursor
objects for updating the database.

In pages.py, use the two distinct endpoints. ... well, actually have
add_issue get called(!)

Also, fix up the coverage testing for the API change.